+++ src/Service/AttrService.php
@@ -4,36 +4,21 @@ declare(strict_types=1);
 
 namespace App\Service;
 
-use App\Entity\Device;
-use App\Model\AttrModel;
-use App\Mapper\AttrMapper;
+use App\Entity\Attr;
 use App\Repository\AttrRepository;
-use App\Repository\DeviceRepository;
-use Doctrine\ORM\EntityManagerInterface;
-use Doctrine\Common\Collections\Collection;
 
 class AttrService
 {
 
     public function __construct(
-        private DeviceRepository $deviceRepo,
         private AttrRepository $attrRepo,
-        private AttrMapper $attrMapper,
-        private EntityManagerInterface $manager
     ) {
 
     }
 
-    public function getAll(): Collection|array
+    public function getAttrById(int $id): Attr
     {
-        $attrModel = $this->attrRepo->findAll();
-        return $this->attrMapper->EntitiesToModels($attrModel);
-    }
-
-    public function getAttr(int $id): AttrModel
-    {
-        $deviceEntity = $this->attrRepo->findOneById($id);
-        return $this->attrMapper->EntityToModel($deviceEntity);
+        return $this->attrRepo->findOneById($id);
     }
 
 }