+++ src/Service/DeviceService.php
@@ -4,26 +4,14 @@ declare(strict_types=1);
 
 namespace App\Service;
 
-use App\Entity\User;
 use App\Entity\Device;
-use App\Model\DevicesModel;
-use App\Mapper\DevicesMapper;
-use App\Model\DeviceModel;
-use App\Mapper\DeviceMapper;
 use App\Repository\DeviceRepository;
-use App\Repository\DevicePropRepository;
-use Doctrine\ORM\EntityManagerInterface;
 
 class DeviceService
 {
     public function __construct(
-        private readonly EntityManagerInterface $manager,
-        private readonly DevicePropRepository $devicePropsRepository,
         private readonly DeviceRepository $deviceRepository,
-        private readonly DeviceMapper $deviceMapper,
-        private readonly DevicesMapper $devicesMapper,
     ) {
-
     }
 
     public function getAll(): Device|array
@@ -31,33 +19,24 @@ class DeviceService
         return $this->deviceRepository->findAll();
     }
 
-    public function getByName(string $name): Device
+    public function getAllId(): array
     {
-        return $this->deviceRepository->findOneByName($name);
+        return $this->deviceRepository->findAllId();
+    }
+
+    public function getAllByPage(int $page, int $limit): array
+    {
+        return $this->deviceRepository->findAllByPage($page, $limit);
     }
 
     public function getById(int $id): Device
     {
-        //return $this->deviceRepository->findDevice($id);
         return $this->deviceRepository->findOneById($id);
     }
 
-    public function getDevices(): array
-    {
-        $deviceEntities = $this->deviceRepository->findAll();
-        return $this->devicesMapper->EntitiesToModels($deviceEntities);
-    }
 
-    public function getDevice(int $id): DeviceModel
+    public function getNumberOfDevices(): int
     {
-        $deviceEntity = $this->deviceRepository->findOneById($id);
-        return $this->deviceMapper->EntityToModel($deviceEntity);
+        return $this->deviceRepository->countAll();
     }
-
-    public function getModelByName(string $name): DevicesModel
-    {
-        $deviceEntity = $this->deviceRepository->findOneByName($name);
-        return $this->devicesMapper->EntityToModel($deviceEntity);
-    }
-
 }