+++ src/Service/PropService.php
@@ -4,34 +4,25 @@ declare(strict_types=1);
 
 namespace App\Service;
 
-use App\Entity\Device;
-use App\Mapper\PropMapper;
-use App\Model\PropModel;
+use App\Entity\Prop;
 use App\Repository\PropRepository;
-use App\Repository\DeviceRepository;
-use Doctrine\ORM\EntityManagerInterface;
 
 class PropService
 {
 
     public function __construct(
-        private DeviceRepository $DeviceRepo,
         private PropRepository $propRepo,
-        private PropMapper $propMapper,
-        private EntityManagerInterface $manager
     ) {
-
     }
 
-    public function getAll(): Device|array
+    public function getAll(): Prop|array
     {
         return $this->propRepo->findAll();
     }
 
-    public function getProps(int $id): PropModel
+    public function getPropById(int $id): Prop
     {
-        $deviceEntity = $this->propRepo->findOneById($id);
-        return $this->propMapper->EntityToModel($deviceEntity);
+        return $this->propRepo->findOneById($id);
     }
 
 }