+++ src/Service/SerializerService.php
@@ -4,12 +4,9 @@ declare(strict_types=1);
 
 namespace App\Service;
 
-use App\Entity\User;
 use Symfony\Component\Serializer\Serializer;
-use Symfony\Component\Serializer\Encoder\XmlEncoder;
 use Symfony\Component\Serializer\Encoder\JsonEncoder;
 use Symfony\Component\Serializer\SerializerInterface;
-use Symfony\Component\Serializer\Normalizer\ObjectNormalizer;
 
 class SerializerService
 {
@@ -18,12 +15,6 @@ class SerializerService
         private SerializerInterface $serializer,
     ) {
 
-    }
-
-    public function json(array $datas): string
-    {
-        $serializer = new Serializer([], [new JsonEncoder()]);
-        return $serializer->encode($datas, 'json');
     }
 
     public function serialize(array|object $datas, array $groups = []): string