+++ src/Service/UserService.php
@@ -6,20 +6,14 @@ namespace App\Service;
 
 use DateTime;
 use App\Entity\User;
-use App\Model\UserModel;
-use App\Mapper\UserMapper;
-use App\Mapper\UsersMapper;
 use App\Service\ClientService;
 use App\Repository\UserRepository;
-use Doctrine\Common\Collections\Collection;
 
 class UserService
 {
 
     public function __construct(
         private UserRepository $userRepo,
-        private UserMapper $userMapper,
-        private UsersMapper $usersMapper,
         private ClientService $clientService,
     ) {
 
@@ -28,18 +22,6 @@ class UserService
     public function getUserById(int $id): User|null
     {
         return $this->userRepo->findOneById($id);
-    }
-
-    public function getUsers(): Collection|array
-    {
-        $usersModel = $this->userRepo->findAll();
-        return $this->usersMapper->EntitiesToModels($usersModel);
-    }
-
-    public function getUser(int $id): UserModel|null
-    {
-        $userModel = $this->userRepo->findOneById($id);
-        return $this->userMapper->EntityToModel($userModel);
     }
 
     public function addUser(User $user, int $clientId): void